requirements.txt
pyqt5==5.15.11
pyodbc==5.2.0
python-dotenv==1.1.0
pandas==2.2.3
openpyxl==3.1.5 